services/orchestrator/requirements.txt
uvloop==0.19.0
asyncpg==0.29.0
nats-py==2.6.0
loguru==0.7.2
